requirements.txt
alembic==1.15.2
blinker==1.9.0
cachelib==0.13.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==45.0.3
defusedxml==0.7.1
Flask==3.1.0
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-OpenID==1.3.1
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.9
MarkupSafe==3.0.2
msgspec==0.19.0
mysqlclient==2.2.7
pycparser==2.22
PyMySQL==1.1.1
python-dotenv==1.1.0
python3-openid==3.2.0
requests==2.32.3
SQLAlchemy==2.0.40
typing_extensions==4.13.1
urllib3==2.3.0
Werkzeug==3.1.3